requirements.txt
elasticsearch==7.17.9
farm-haystack==1.14.0
fastapi==0.94.1
gensim==3.8.3
huggingface-hub==0.13.2
matplotlib==3.7.1
nltk==3.8.1
numpy==1.23.5
pandas==1.5.3
pillow==9.4.0
scikit-learn==1.2.2
scipy==1.10.1
spacy==2.3.9
texthero==1.1.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
transformers==4.25.1